+++ WatchUI/keywords/__init__.py
@@ -1,0 +1,3 @@
+from _visual import _VisualKeywords
+
+__all__ = ["_VisualKeywords"]
